Transforms and matrices

For basic transformations of the drawing space, use scale(sx, sy), rotate(a), and translate(tx, ty).

Base.scale — Function.

Scale subsequent drawing in x and y.

Example:

scale(0.2, 0.3)

source

Luxor.rotate — Function.

Rotate subsequent drawing by a radians clockwise.

rotate(a)

source

Luxor.translate — Function.

Translate to new location.

translate(x, y)

or

translate(point)

source

The current matrix is a six element array, perhaps like this:

[1, 0, 0, 1, 0, 0]

getmatrix() gets the current matrix, setmatrix(a) sets the matrix to array a, and transform(a) transforms the current matrix by 'multiplying' it with matrix a.

Luxor.getmatrix — Function.

Get the current matrix.

getmatrix()

Return current Cairo matrix as an array. In Cairo and Luxor, a matrix is an array of 6 float64 numbers:

xx component of the affine transformation

yx component of the affine transformation

xy component of the affine transformation

yy component of the affine transformation

x0 translation component of the affine transformation

y0 translation component of the affine transformation

Some basic matrix transforms:

translate(dx,dy) = transform([1, 0, 0, 1, dx, dy]) shift by

scale(fx, fy) = transform([fx, 0, 0, fy, 0, 0]) scale by

rotate(A) = transform([c, s, -c, c, 0, 0]) rotate to A radians

x-skew(a) = transform([1, 0, tan(a), 1, 0, 0]) xskew by A

y-skew(a) = transform([1, tan(a), 0, 1, 0, 0]) yskew by A

flip HV = transform([fx, 0, 0, fy, cx(1-fx), cy (fy-1)]) flip

When a drawing is first created, the matrix looks like this:

getmatrix() = [1.0, 0.0, 0.0, 1.0, 0.0, 0.0]

When the origin is moved to 400/400, it looks like this:

getmatrix() = [1.0, 0.0, 0.0, 1.0, 400.0, 400.0]

To reset the matrix to the original:

setmatrix([1.0, 0.0, 0.0, 1.0, 0.0, 0.0])

source

Luxor.setmatrix — Function.

Change the current Cairo matrix to matrix m.

setmatrix(m::Array)

Use getmatrix() to get the current matrix.

source

Luxor.transform — Function.

Modify the current matrix by multiplying it by matrix a.

transform(a::Array)

For example, to skew the current state by 45 degrees in x and move by 20 in y direction:

transform([1, 0, tand(45), 1, 0, 20])

Use getmatrix() to get the current matrix.

source